requirements.txt
asgiref==3.3.4
Django==3.2
django-crispy-forms==1.11.2
djangorestframework==3.12.4
flake8==3.9.1
importlib-metadata==4.0.1
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.1
